Made Reusable Emoji component
