+++ src/components/index.js
@@ -11,6 +11,7 @@ import PostEditor from "./PostEditor";
 import PostActions from "./PostActions";
 import CommentsAndShares from "./CommentsAndShares";
 import PeopleListModal from "./PeopleListModal";
+import EmojiPicker from "./EmojiPicker";
 
 export {
   SearchBar,
@@ -26,4 +27,5 @@ export {
   PostActions,
   CommentsAndShares,
   PeopleListModal,
+  EmojiPicker,
 };
